fix business show, add efficiency to review/business fetch

+++ frontend/src/components/BusinessItemShow/index.js
@@ -13,16 +13,17 @@ import styles from './BusinessItemShow.module.css'
 
 const BusinessItemShow = () => {
 
-    const sessionUser = useSelector(state => state.session.user)
+    let sessionUser = useSelector(state => state.session.user)
+    sessionUser = sessionUser ? sessionUser : undefined
     const allUsers = useSelector(getUsers)
     const dispatch = useDispatch();
     const businessParam = useParams()
     const [reviewBody, setReviewBody] = useState('')
     const business = useSelector(getBusiness(businessParam.id))
-    const reviews = useSelector(getReviews)
+    const reviews = useSelector(getReviews(businessParam.id))
     const [reviewCount, setReviewCount] = useState(0)
     const [reviewsArray, setReviewsArray] = useState([])
-
+    console.log(reviews);
     useEffect(() => {
         if (!business) {
             dispatch(fetchBusiness(businessParam.id))
@@ -30,47 +31,47 @@ const BusinessItemShow = () => {
     }, [businessParam.id])
 
     useEffect(() => {
-        dispatch(fetchReviews())
-        dispatch(fetchUsers())
+        dispatch(fetchUsers()).then(() => dispatch(fetchReviews()))
     }, [dispatch] )
 
-    useEffect(() => {
-        const mappedReviews = reviews.map((review) => {
-            let reviewUserName = ''
-            if (parseInt(businessParam.id) === review.businessId) {
-                allUsers.forEach((user) => {
-                    if (user.id === review.userId) {
-                        reviewUserName = (`${user.first_name} ${user.last_name}`)
-                    };
-                })
-                let deleteAndUpdateButtons = ''
-                if (review.userId === sessionUser.id) {
-                    deleteAndUpdateButtons = ( 
-                    <>
-                        <button className={styles.deleteReview} onClick={(e) => dispatch(destroyReview(review.id))} key={review.id}> Delete </button>
-                        <button className={styles.editReview} onClick={(e) => dispatch(updateReview(review))} key={review.id}> Update </button>
-                    </>
-                    )
-                }
-                
-                return (
-                    <div key={review.id} className={styles.singleReview}>
-                        <svg className={styles.reviewMenu}>
-                        <path d="M12 13.5a1.5 1.5 0 100-3 1.5 1.5 0 000 3zm8 0a1.5 1.5 0 100-3 1.5 1.5 0 000 3zm-16 0a1.5 1.5 0 100-3 1.5 1.5 0 000 3z"></path>
-                        </svg>
-                        <p className={styles.reviewerName}>{reviewUserName}</p>
-                        <p className={styles.cityState}> New York, NY </p>
-                        <br />
-                        <p> Rating: 4/5 | {review.created_at} </p>
-                        <br />
-                        <p review={review} className={styles.paragraph}> {review.body} </p>
-                        {deleteAndUpdateButtons}
-                    </div>
-                )
-            }
-        });
-        setReviewsArray(mappedReviews)
-    }, [])
+    if (!business) return null;
+    // useEffect(() => {
+    // }, [])
+    const mappedReviews = reviews.map((review) => {
+        let reviewUserName = ''
+
+        allUsers.forEach((user) => {
+            if (user.id === review.userId) {
+                reviewUserName = (`${user.first_name} ${user.last_name}`)
+            };
+        })
+        let deleteAndUpdateButtons = ''
+        if (sessionUser && review.userId === sessionUser.id) {
+            deleteAndUpdateButtons = ( 
+            <>
+                <button className={styles.deleteReview} onClick={(e) => dispatch(destroyReview(review.id))} key={review.id}> Delete </button>
+                <button className={styles.editReview} onClick={(e) => dispatch(updateReview(review))} key={review.id}> Update </button>
+            </>
+            )
+        }
+        
+        return (
+            <div key={review.id} className={styles.singleReview}>
+                <svg className={styles.reviewMenu}>
+                <path d="M12 13.5a1.5 1.5 0 100-3 1.5 1.5 0 000 3zm8 0a1.5 1.5 0 100-3 1.5 1.5 0 000 3zm-16 0a1.5 1.5 0 100-3 1.5 1.5 0 000 3z"></path>
+                </svg>
+                <p className={styles.reviewerName}>{reviewUserName}</p>
+                <p className={styles.cityState}> New York, NY </p>
+                <br />
+                <p> Rating: 4/5 | {review.created_at} </p>
+                <br />
+                <p review={review} className={styles.paragraph}> {review.body} </p>
+                {deleteAndUpdateButtons}
+            </div>
+        )
+    });
+
+    // setReviewsArray(mappedReviews)
 
     const handleSubmit = (e) => {
         const reviewObject = {body: reviewBody, user_id: sessionUser.id, business_id: businessParam.id}
@@ -116,7 +117,7 @@ const BusinessItemShow = () => {
                         <path fill="white" fill-rule="evenodd" clip-rule="evenodd" d="M16 21.3978L20.1518 23.5378C20.3804 23.6557 20.6409 23.697 20.8949 23.6555C21.5489 23.5488 21.9926 22.932 21.8859 22.2779L21.1336 17.668L24.4519 14.3807C24.6346 14.1997 24.7544 13.9647 24.7934 13.7104C24.894 13.0553 24.4445 12.4427 23.7895 12.3421L19.1727 11.6331L17.0717 7.4614C16.956 7.23163 16.7695 7.04513 16.5397 6.92941C15.9478 6.63131 15.2263 6.86949 14.9282 7.4614L12.8272 11.6331L8.21047 12.3421C7.95619 12.3812 7.72118 12.5009 7.54013 12.6837C7.0737 13.1545 7.07727 13.9143 7.54809 14.3807L10.8664 17.668L10.114 22.2779C10.0726 22.5318 10.1139 22.7923 10.2317 23.021C10.5354 23.6101 11.2591 23.8415 11.8482 23.5378L16 21.3978V21.3978Z"></path>
                     </svg>
                     <p id={styles.numOfReviews}>
-                        {reviewsArray.length} reviews
+                        {mappedReviews.length} reviews
                     </p>
                 </p>
             </div>
@@ -127,7 +128,7 @@ const BusinessItemShow = () => {
             </form>
             <div className={styles.reviews}>
                 <h3 id={styles.reviewHeader}>Recommended Reviews</h3>
-                {reviewsArray}
+                {mappedReviews}
             </div>
